# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/b": {
            "get": {
                "description": "Branch to another API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branch"
                ],
                "summary": "Branch",
                "parameters": [
                    {
                        "type": "string",
                        "example": "123",
                        "name": "branch_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Details about the client",
                        "name": "client_detail",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "456",
                        "name": "requestID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "UTC+0",
                        "name": "timezone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1.0.2",
                        "name": "wn_ios_version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BrResponse"
                        }
                    }
                }
            }
        },
        "/p": {
            "post": {
                "description": "Payment to another API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Payment",
                "parameters": [
                    {
                        "description": "Payment Request",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BrResponse": {
            "type": "object",
            "properties": {
                "employee_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Employee"
                    }
                },
                "permission_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Permission"
                    }
                },
                "requestID": {
                    "type": "string",
                    "example": "REQ-001"
                },
                "success": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "dto.Channel": {
            "type": "object",
            "properties": {
                "channel_order_no": {
                    "type": "string",
                    "example": "456"
                },
                "order_channel_id": {
                    "type": "string",
                    "example": "123"
                }
            }
        },
        "dto.Choice": {
            "type": "object",
            "properties": {
                "c_id": {
                    "type": "string",
                    "example": "C-001"
                },
                "count": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "dto.CommentObj": {
            "type": "object",
            "properties": {
                "company_address": {
                    "type": "string",
                    "example": "456 Business Ave, City, Country"
                },
                "company_branch": {
                    "type": "string",
                    "example": "Branch 1"
                },
                "company_country_code": {
                    "type": "string",
                    "example": "US"
                },
                "company_email": {
                    "type": "string",
                    "example": "info@company.com"
                },
                "company_name": {
                    "type": "string",
                    "example": "Company Inc."
                },
                "company_phone": {
                    "type": "string",
                    "example": "0987654321"
                },
                "company_prefix": {
                    "type": "string",
                    "example": "COMP"
                },
                "company_tax_id": {
                    "type": "string",
                    "example": "TAX123"
                }
            }
        },
        "dto.Employee": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "end_time": {
                    "type": "string",
                    "example": "2023-07-03T15:04:05Z07:00"
                },
                "firstname": {
                    "type": "string",
                    "example": "John"
                },
                "image_key": {
                    "type": "string",
                    "example": "img123456"
                },
                "lastname": {
                    "type": "string",
                    "example": "Doe"
                },
                "owner_class_id": {
                    "type": "string",
                    "example": "001"
                },
                "owner_class_name": {
                    "type": "string",
                    "example": "VIP"
                },
                "owner_id": {
                    "type": "string",
                    "example": "1"
                },
                "phonenumber": {
                    "type": "string",
                    "example": "555-1234"
                },
                "start_time": {
                    "type": "string",
                    "example": "2023-07-01T15:04:05Z07:00"
                },
                "under_id": {
                    "type": "string",
                    "example": "2"
                }
            }
        },
        "dto.MenuGroupID": {
            "type": "object",
            "properties": {
                "cat_grp_id": {
                    "type": "string",
                    "example": "CAT-001"
                },
                "cat_grp_name": {
                    "type": "string",
                    "example": "Category1"
                }
            }
        },
        "dto.MenuID": {
            "type": "object",
            "properties": {
                "cost_center": {
                    "type": "string",
                    "example": "CC-001"
                },
                "dept_code": {
                    "type": "string",
                    "example": "DPT-001"
                }
            }
        },
        "dto.MposCreditDetail": {
            "type": "object",
            "properties": {
                "approvalCode": {
                    "type": "string",
                    "example": "APPROVED"
                },
                "cardHolderName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "cardNo": {
                    "type": "string",
                    "example": "4111111111111111"
                },
                "cardType": {
                    "type": "string",
                    "example": "VISA"
                },
                "reference": {
                    "type": "string",
                    "example": "REF-001"
                },
                "statusCode": {
                    "type": "string",
                    "example": "200"
                },
                "statusMessage": {
                    "type": "string",
                    "example": "Approved"
                },
                "traceNo": {
                    "type": "string",
                    "example": "TRACE-001"
                },
                "transactionID": {
                    "type": "string",
                    "example": "TXN-001"
                }
            }
        },
        "dto.Option": {
            "type": "object",
            "properties": {
                "action_by": {
                    "type": "string",
                    "example": "Chef1"
                },
                "payment_item_id": {
                    "type": "string",
                    "example": "PI-001"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1
                },
                "void_flag": {
                    "type": "boolean",
                    "example": false
                },
                "void_reason": {
                    "type": "string",
                    "example": "None"
                }
            }
        },
        "dto.OrderCustomerDetail": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "123 Street Name, City, Country"
                },
                "comment": {
                    "type": "string",
                    "example": "No onions please"
                },
                "country_code": {
                    "type": "string",
                    "example": "US"
                },
                "cust_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "phone_number": {
                    "type": "string",
                    "example": "1234567890"
                },
                "phone_prefix": {
                    "type": "string",
                    "example": "+1"
                }
            }
        },
        "dto.OrderMenu": {
            "type": "object",
            "properties": {
                "action_by": {
                    "type": "string",
                    "example": "Chef1"
                },
                "cancel_flag": {
                    "type": "boolean",
                    "example": false
                },
                "is_customer": {
                    "type": "boolean",
                    "example": true
                },
                "menu_comment": {
                    "type": "string",
                    "example": "Extra crispy"
                },
                "menu_group_id": {
                    "type": "string",
                    "example": "MG-001"
                },
                "menu_group_name": {
                    "type": "string",
                    "example": "Appetizers"
                },
                "menu_id": {
                    "type": "string",
                    "example": "MENU-001"
                },
                "menu_name": {
                    "type": "string",
                    "example": "Spring Rolls"
                },
                "menu_order_time": {
                    "type": "string",
                    "example": "2024-07-03T12:05:00Z"
                },
                "menu_price": {
                    "type": "number",
                    "example": 5
                },
                "option_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Option"
                    }
                },
                "order_menu_id": {
                    "type": "string",
                    "example": "OM-001"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "reject_by": {
                    "type": "string",
                    "example": "Chef2"
                },
                "reject_flag": {
                    "type": "boolean",
                    "example": false
                },
                "reject_reason": {
                    "type": "string",
                    "example": "Out of stock"
                },
                "reject_time": {
                    "type": "string",
                    "example": "2024-07-03T12:10:00Z"
                },
                "send_order_by": {
                    "type": "string",
                    "example": "Chef1"
                },
                "send_order_system_id": {
                    "type": "string",
                    "example": "SYS-001"
                },
                "served_by": {
                    "type": "string",
                    "example": "Server1"
                },
                "served_time": {
                    "type": "string",
                    "example": "2024-07-03T12:15:00Z"
                }
            }
        },
        "dto.PaRequest": {
            "type": "object",
            "properties": {
                "bank_code": {
                    "type": "string",
                    "example": "XYZ"
                },
                "before_svc": {
                    "type": "number",
                    "example": 100
                },
                "before_vat": {
                    "type": "number",
                    "example": 93
                },
                "bill_discount_by": {
                    "type": "string",
                    "example": "Admin"
                },
                "bill_discount_promotion_id": {
                    "type": "string",
                    "example": "BDP-001"
                },
                "bill_discount_reason": {
                    "type": "string",
                    "example": "Seasonal Discount"
                },
                "bill_discount_type": {
                    "type": "string",
                    "example": "Percentage"
                },
                "bill_discount_value": {
                    "type": "number",
                    "example": 10
                },
                "bill_discounted_price": {
                    "type": "number",
                    "example": 90
                },
                "branch_id": {
                    "type": "string",
                    "example": "BR-001"
                },
                "branch_uuid": {
                    "type": "string",
                    "example": "BR-UUID-001"
                },
                "buffet_package_id": {
                    "type": "string",
                    "example": "BP-001"
                },
                "cancel_bill": {
                    "type": "boolean",
                    "example": false
                },
                "cancel_menu_flag": {
                    "type": "boolean",
                    "example": false
                },
                "cardNo": {
                    "type": "string",
                    "example": "1111"
                },
                "cash_drawer_code": {
                    "type": "string",
                    "example": "CD-001"
                },
                "change_amount": {
                    "type": "number",
                    "example": 0
                },
                "channel": {
                    "$ref": "#/definitions/dto.Channel"
                },
                "client_ip": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "client_name": {
                    "type": "string",
                    "example": "Client1"
                },
                "close_owner_id": {
                    "type": "string",
                    "example": "OWNER-002"
                },
                "comment_obj": {
                    "$ref": "#/definitions/dto.CommentObj"
                },
                "country_code": {
                    "type": "string",
                    "example": "US"
                },
                "create_dt": {
                    "type": "string",
                    "example": "2024-07-03T12:00:00Z"
                },
                "credit_card_id": {
                    "type": "string",
                    "example": "CC-001"
                },
                "crmRedeemId": {
                    "type": "string",
                    "example": "CRM-REDEEM-001"
                },
                "crm_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "crm_phone_no": {
                    "type": "string",
                    "example": "1234567890"
                },
                "crm_system_id": {
                    "type": "string",
                    "example": "CRM-SYS-001"
                },
                "cust_pmt_id": {
                    "type": "string",
                    "example": "CPMT-001"
                },
                "cust_pmt_ref": {
                    "type": "string",
                    "example": "CPMT-REF-001"
                },
                "customer_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "discount_amount": {
                    "type": "number",
                    "example": 10
                },
                "edc_serial_no": {
                    "type": "string",
                    "example": "789"
                },
                "fs_crm_history_id": {
                    "type": "string",
                    "example": "FS-CRM-HIST-001"
                },
                "fs_crm_member_id": {
                    "type": "string",
                    "example": "FS-CRM-MEM-001"
                },
                "fs_crm_promotion_uuid": {
                    "type": "string",
                    "example": "FS-CRM-PROMO-UUID-001"
                },
                "fs_redeem_id": {
                    "type": "string",
                    "example": "FS-REDEEM-001"
                },
                "host_name": {
                    "type": "string",
                    "example": "localhost"
                },
                "kiosk_type": {
                    "type": "string",
                    "example": "Self-Service"
                },
                "latitude": {
                    "type": "number",
                    "example": 40.7128
                },
                "longitude": {
                    "type": "number",
                    "example": -74.006
                },
                "memberId": {
                    "type": "string",
                    "example": "MEM-001"
                },
                "mpos_credit_detail": {
                    "$ref": "#/definitions/dto.MposCreditDetail"
                },
                "no_of_menu": {
                    "type": "integer",
                    "example": 2
                },
                "non_vat_amount": {
                    "type": "number",
                    "example": 0
                },
                "open_order_dt": {
                    "type": "string",
                    "example": "2024-07-03T12:00:00Z"
                },
                "open_owner_id": {
                    "type": "string",
                    "example": "OWNER-001"
                },
                "order_customer_detail": {
                    "$ref": "#/definitions/dto.OrderCustomerDetail"
                },
                "order_id": {
                    "type": "string",
                    "example": "ORDER-001"
                },
                "order_menu_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OrderMenu"
                    }
                },
                "order_type": {
                    "type": "string",
                    "example": "Dine-In"
                },
                "payment_amount": {
                    "type": "number",
                    "example": 90
                },
                "payment_channel": {
                    "type": "string",
                    "example": "Credit Card"
                },
                "payment_id": {
                    "type": "string",
                    "example": "PMT-001"
                },
                "payment_type": {
                    "type": "string",
                    "example": "Credit Card"
                },
                "phone_number": {
                    "type": "string",
                    "example": "1234567890"
                },
                "phone_prefix": {
                    "type": "string",
                    "example": "+1"
                },
                "pmt_api_ref": {
                    "type": "string",
                    "example": "PMT-API-001"
                },
                "pmt_menu_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PmtMenu"
                    }
                },
                "pos_id": {
                    "type": "string",
                    "example": "POS-001"
                },
                "prefix_cardNo": {
                    "type": "string",
                    "example": "4111"
                },
                "pro_user_id": {
                    "type": "string",
                    "example": "PRO-001"
                },
                "promotion_name": {
                    "type": "string",
                    "example": "Summer Sale"
                },
                "promotion_ref_code": {
                    "type": "string",
                    "example": "PROMO-001"
                },
                "qr_text": {
                    "type": "string",
                    "example": "QR-001"
                },
                "question_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Question"
                    }
                },
                "receipt_no": {
                    "type": "string",
                    "example": "REC-001"
                },
                "refund_amount": {
                    "type": "number",
                    "example": 0
                },
                "requestID": {
                    "type": "string",
                    "example": "REQ-001"
                },
                "restaurant_id": {
                    "type": "string",
                    "example": "RES-001"
                },
                "resv_id": {
                    "type": "string",
                    "example": "RESV-001"
                },
                "rounding_amount": {
                    "type": "number",
                    "example": 0.05
                },
                "rounding_type": {
                    "type": "string",
                    "example": "Up"
                },
                "running_book": {
                    "type": "string",
                    "example": "RB-001"
                },
                "running_no": {
                    "type": "string",
                    "example": "RN-001"
                },
                "seat_amt": {
                    "type": "integer",
                    "example": 4
                },
                "service": {
                    "type": "number",
                    "example": 10
                },
                "service_percent": {
                    "type": "number",
                    "example": 10
                },
                "split_payment_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SplitPaymentDetail"
                    }
                },
                "sub_amount": {
                    "type": "number",
                    "example": 100
                },
                "table_id": {
                    "type": "string",
                    "example": "TBL-001"
                },
                "table_name": {
                    "type": "string",
                    "example": "Table 5"
                },
                "tax": {
                    "type": "number",
                    "example": 7
                },
                "tax_percent": {
                    "type": "number",
                    "example": 7
                },
                "the_one_card_code": {
                    "type": "string",
                    "example": "ONE-001"
                },
                "timezone": {
                    "type": "string",
                    "example": "GMT+7"
                },
                "tips_amount": {
                    "type": "number",
                    "example": 5
                },
                "total_amount": {
                    "type": "number",
                    "example": 100
                },
                "unique_order_code": {
                    "type": "string",
                    "example": "UNIQUE-123"
                },
                "version": {
                    "type": "string",
                    "example": "1.0"
                },
                "void_after_print_inv": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.PaResponse": {
            "type": "object",
            "properties": {
                "flow_acct_resp": {
                    "type": "string",
                    "example": "Tax invoice response here"
                },
                "freshLivingResp": {
                    "type": "string",
                    "example": "No Data"
                },
                "is_void": {
                    "type": "boolean",
                    "example": false
                },
                "member_id": {
                    "type": "string",
                    "example": "MEM-004"
                },
                "order_id": {
                    "type": "string",
                    "example": "ORDER-001"
                },
                "payment_id": {
                    "type": "string",
                    "example": "PAY-123"
                },
                "payment_ref_id": {
                    "type": "string",
                    "example": "REF-002"
                },
                "pmt_time": {
                    "type": "string",
                    "example": "2024-07-03T12:00:00Z"
                },
                "qr_text": {
                    "type": "string",
                    "example": "HEADER_QR-QR-789"
                },
                "receipt_no": {
                    "type": "string",
                    "example": "REC-456"
                },
                "requestID": {
                    "type": "string",
                    "example": "REQ-001"
                },
                "res_mem_id": {
                    "type": "string",
                    "example": "MEM-003"
                },
                "review_point": {
                    "type": "integer",
                    "example": 5
                },
                "split_pmt_detail": {
                    "type": "string",
                    "example": "Split payment details here"
                },
                "success": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "dto.Permission": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string",
                    "example": "permission 1"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "dto.PmtMenu": {
            "type": "object",
            "properties": {
                "action_by": {
                    "type": "string",
                    "example": "Admin"
                },
                "discount_reason": {
                    "type": "string",
                    "example": "Loyalty Discount"
                },
                "discount_type": {
                    "type": "string",
                    "example": "Percentage"
                },
                "discount_value": {
                    "type": "number",
                    "example": 1
                },
                "discounted_price": {
                    "type": "number",
                    "example": 9
                },
                "fs_crm_promotion_uuid": {
                    "type": "string",
                    "example": "FS-CRM-PROMO-UUID-002"
                },
                "is_non_vat": {
                    "type": "boolean",
                    "example": false
                },
                "menu_group_id": {
                    "$ref": "#/definitions/dto.MenuGroupID"
                },
                "menu_group_name": {
                    "type": "string",
                    "example": "Main Course"
                },
                "menu_id": {
                    "$ref": "#/definitions/dto.MenuID"
                },
                "menu_name": {
                    "type": "string",
                    "example": "Pizza"
                },
                "option_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Option"
                    }
                },
                "order_menu_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OrderMenu"
                    }
                },
                "override_promotion": {
                    "type": "boolean",
                    "example": true
                },
                "payment_item_id": {
                    "type": "string",
                    "example": "PMT-001"
                },
                "price": {
                    "type": "number",
                    "example": 10
                },
                "price_with_option": {
                    "type": "number",
                    "example": 10
                },
                "pro_user_id": {
                    "type": "string",
                    "example": "PRO-002"
                },
                "promotion_id": {
                    "type": "string",
                    "example": "PROMO-002"
                },
                "promotion_menu_group_id": {
                    "type": "string",
                    "example": "PMG-001"
                },
                "promotion_menu_id": {
                    "type": "string",
                    "example": "PROMO-001"
                },
                "promotion_name": {
                    "type": "string",
                    "example": "Holiday Discount"
                },
                "promotion_ref_code": {
                    "type": "string",
                    "example": "PROMO-REF-001"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1
                },
                "void_flag": {
                    "type": "boolean",
                    "example": false
                },
                "void_reason": {
                    "type": "string",
                    "example": "None"
                }
            }
        },
        "dto.Question": {
            "type": "object",
            "properties": {
                "choice_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Choice"
                    }
                },
                "q_id": {
                    "type": "string",
                    "example": "Q-001"
                }
            }
        },
        "dto.SplitPaymentDetail": {
            "type": "object",
            "properties": {
                "card_digit": {
                    "type": "string",
                    "example": "5678"
                },
                "change_amount": {
                    "type": "number",
                    "example": 10
                },
                "credit_card_id": {
                    "type": "string",
                    "example": "CC-002"
                },
                "cust_pmt_ref": {
                    "type": "string",
                    "example": "CPMT-REF-002"
                },
                "paid_amount": {
                    "type": "number",
                    "example": 50
                },
                "payment_type": {
                    "type": "string",
                    "example": "Cash"
                },
                "prefix_card_digit": {
                    "type": "string",
                    "example": "1234"
                },
                "split_amount": {
                    "type": "number",
                    "example": 40
                }
            }
        }
    }
}